+++ DesktopApplicationTemplate.UI/Views/MainWindow.xaml.cs
@@ -101,7 +101,11 @@ namespace DesktopApplicationTemplate.UI.Views
                 {
                     if (vm.Logger != null)
                     {
-                        logger.LogAdded += entry => svc.AddLog(entry.Message, entry.Color, entry.Level);
+                        logger.LogAdded += entry =>
+                        {
+                            var brush = (Brush?)new BrushConverter().ConvertFromString(entry.Color);
+                            svc.AddLog(entry.Message, brush, entry.Level);
+                        };
                     }
                 }